requirements.txt
# Production-ready requirements for 51Talk AI Learning Platform
# Core Flask dependencies
Flask==2.3.3
Flask-Mail==0.9.1
Werkzeug==2.3.7
Jinja2==3.1.2

# Database
psycopg2-binary==2.9.7

# Configuration
python-dotenv==1.0.0

# Document Processing & AI
langchain==0.1.14
langchain-community==0.0.38
python-pptx==0.6.21
pypdf==3.17.4

# Vector Store & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# HTTP requests for HF API
requests==2.31.0

# Utilities
pytz==2024.1
typing-extensions==4.10.0

# Production server
gunicorn==21.2.0

# Security (for production)
cryptography==41.0.7